TINY-10370: create separate folder for InserContent related tests
